+++ pkg/core/common/accounts_test.go
@@ -1,4 +1,4 @@
-package accounts
+package common
 
 import (
 	"testing"
